backend/requirements.txt
Django
djangorestframework
python-decouple
gunicorn
whitenoise
PyJWT